add item to bundle

+++ backend/models/bundles.go
@@ -1,7 +1,7 @@
 package models
 
 type Bundle struct {
-	Name        string            `json:"name"`
-	Description string            `json:"description"`
-	Items       map[string]uint16 `json:"items"`
+	Name        string         `json:"name"`
+	Description string         `json:"description"`
+	Items       map[string]int `json:"items"`
 }
